refactor story types
