+++ src/components/Generic/ProjectCard/DraggableProjectCard.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react'
 import DraggableProjectCard from './DraggableProjectCard'
 import { ProjectCardType } from './DraggableProjectCard'
-import { mockProjects } from '@/mocks/Projects.mock'
+import { mockProjects } from '@/mocks/newProjects.mock'
 
 const meta: Meta<typeof DraggableProjectCard> = {
   component: DraggableProjectCard,
